Refactor auth steps in guardian script
